+++ samples/elaborate/contact_manager/contact_manager_presenter.rb
@@ -1,5 +1,5 @@
 # Copyright (c) 2007-2022 Andy Maleh
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
 # "Software"), to deal in the Software without restriction, including
@@ -7,10 +7,10 @@
 # distribute, sublicense, and/or sell copies of the Software, and to
 # permit persons to whom the Software is furnished to do so, subject to
 # the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 # MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -33,15 +33,37 @@ class ContactManager
     end
   
     def list
-      self.results = @contact_repository.find({})
+      @applied_filter = {}
+      filter
     end
   
     def find
-      filter_map = {}
+      @applied_filter = fields
+      filter
+    end
+    
+    def filter
+      self.results = @contact_repository.find(@applied_filter)
+    end
+    
+    def fields
+      filters = {}
       @@contact_attributes.each do |attribute_name|
-        filter_map[attribute_name] = self.send(attribute_name) if self.send(attribute_name)
+        filters[attribute_name] = self.send(attribute_name) if self.send(attribute_name)
       end
-      self.results = @contact_repository.find(filter_map)
+      filters
+    end
+    
+    def clear_fields
+      @@contact_attributes.each do |attribute_name|
+        self.send("#{attribute_name}=", '')
+      end
+    end
+    
+    def create
+      @contact_repository.create(fields)
+      filter
+      clear_fields
     end
   end
 end